+++ source/extensions/omni.isaac.app.setup/omni/isaac/app/setup/extension.py
@@ -286,11 +286,6 @@ class CreateSetupExtension(omni.ext.IExt):
         return None
 
     def __menu_update(self):
-        self._viewport = None
-        try:
-            self._viewport = omni.kit.viewport_legacy.get_viewport_interface()
-        except:
-            pass
         self._appwindow = omni.appwindow.get_default_app_window()
         self._settings = carb.settings.get_settings()
         self._selection = omni.usd.get_context().get_selection()