+++ source/extensions/omni.isaac.core_nodes/nodes/OgnIsaacSetViewportResolution.py
@@ -7,7 +7,8 @@
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 
 import omni
-import omni.kit.viewport_legacy
+from omni.kit.viewport.utility import get_viewport_from_window_name, get_active_viewport
+
 from omni.isaac.core.utils.carb import set_carb_setting
 import carb
 
@@ -19,14 +20,16 @@ class OgnIsaacSetViewportResolution:
 
     @staticmethod
     def compute(db) -> bool:
-        vp = omni.kit.viewport_legacy.get_viewport_interface()
-        for instance in vp.get_instance_list():
-            if vp.get_viewport_window_name(instance) == db.inputs.viewport:
-                window = vp.get_viewport_window(instance)
-                window.set_texture_resolution(db.inputs.width, db.inputs.height)
-                set_carb_setting(carb.settings.get_settings(), "/app/hydra/aperture/conform", 3)
-                set_carb_setting(carb.settings.get_settings(), "/app/hydra/aperture/conform", 4)
-                db.outputs.execOut = omni.graph.core.ExecutionAttributeState.ENABLED
-                return True
+        viewport_name = db.inputs.viewport
+        if viewport_name:
+            viewport_api = get_viewport_from_window_name(viewport_name)
+        else:
+            viewport_api = get_active_viewport()
+
+        if viewport_api:
+            viewport_api.set_texture_resolution((db.inputs.width, db.inputs.height))
+            set_carb_setting(carb.settings.get_settings(), "/app/hydra/aperture/conform", 3)
+            set_carb_setting(carb.settings.get_settings(), "/app/hydra/aperture/conform", 4)
+
         db.outputs.execOut = omni.graph.core.ExecutionAttributeState.ENABLED
         return True