+++ source/extensions/omni.isaac.utils/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.2.0] - 2022-09-01
+
+### Changed
+- removed legacy viewport calls
 ## [0.1.11] - 2022-03-16
 
 ### Changed